Layout Model Refactor
