--- packages/layout/src/tree-layout/layout-definition/tree/Node.js
@@ -1,142 +1,0 @@
-import { getNodeId } from '../utils';
-
-class Node {
-    constructor (data) {
-        this._model = data;
-        this._parent = null;
-        this._children = [];
-        this._parentCut = null;
-        this._boundBox = {
-            top: null,
-            left: null,
-            height: null,
-            width: null
-        };
-
-        this._id = getNodeId();
-    }
-
-    addChildren (entries) {
-        this._children.push(...entries);
-        entries.forEach((e) => { e.parent(this); });
-    }
-
-    isRoot () {
-        return this._parent === null;
-    }
-
-    isLeaf () {
-        return !this._children.length;
-    }
-
-    getCutType () {
-        return this._model.cut();
-    }
-
-    isPreferred () {
-        return !!this._model.preferred();
-    }
-
-    children (children) {
-        if (children) {
-            this._children = children;
-        }
-        return this._children;
-    }
-
-    parent (parent) {
-        if (parent) {
-            this._parent = parent;
-        }
-        return this._parent;
-    }
-
-    parentCut (parentCut) {
-        if (parentCut) {
-            this._parentCut = parentCut;
-        }
-        return this._parentCut;
-    }
-
-    id () {
-        return this._id;
-    }
-
-    model (model) {
-        if (model) {
-            this._model = model;
-        }
-        return this._model;
-    }
-
-    boundBox (bound) {
-        if (bound) {
-            this._boundBox = {
-                top: bound.top,
-                left: bound.left,
-                height: bound.height,
-                width: bound.width
-            };
-        }
-        return this._boundBox;
-    }
-  /**
-     * function to search a node and update it with the config provided.
-     * @param  {Object} nodeconfig
-     */
-    updateNode (nodeconfig) {
-        if (this._id === nodeconfig._id) {
-            this._model.cut(nodeconfig.cut);
-            this._model.ratioWeight(nodeconfig.ratioWeight);
-        } else {
-            this._children.forEach((node) => {
-                if (node._id === nodeconfig._id) {
-                    node.model.cut(nodeconfig.cut);
-                    node.model.ratioWeight(nodeconfig.ratioWeight);
-                    return;
-                }
-                this.searchNode(node, nodeconfig);
-            });
-        }
-    }
-
-  // Recursive function to search a node
-    searchNode (node, nodeconfig) {
-        node.children().forEach((node1) => {
-            if (node1.id() === nodeconfig._id) {
-                node1.model().cut(nodeconfig.cut);
-                node1.model().ratioWeight(nodeconfig.ratioWeight);
-            } else {
-                this.searchNode(node1, nodeconfig);
-            }
-        });
-    }
-
-  /**
-     * function to delete a node from tree Structure.
-     * @param  {String} nodeId - node Id of the Node
-     */
-    delete (nodeId) {
-        this._children.forEach((node) => {
-            if (node.id() === nodeId) {
-                const index = this._children.indexOf(node);
-                this._model.lanes().splice(index, 1);
-            }
-            this.deleteSearchNode(node, nodeId);
-        });
-    }
-
-  // Recursive function to search a node
-    deleteSearchNode (node, nodeId) {
-        node.children().forEach((node1) => {
-            if (node1.id() === nodeId) {
-                const index = node.children().indexOf(node1);
-                node.model().lanes().splice(index, 1);
-            } else {
-                this.deleteSearchNode(node1, nodeId);
-            }
-        });
-    }
-}
-
-export default Node;
