--- packages/layout/src/tree-layout/layout-definition/utils/index.js
@@ -1,15 +1,0 @@
-import {
-  isEqual,
-  getNodeId,
-  yExtraSpace,
-  xExtraSpace,
-  determineBoundBox
-} from './utils';
-
-export {
-  isEqual,
-  getNodeId,
-  yExtraSpace,
-  xExtraSpace,
-  determineBoundBox
-};
